pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "bunq-nest-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx[http2]>=0.27",
    "anthropic>=0.40",
    "sse-starlette>=1.8",
    "python-multipart>=0.0.9",
    "pyjwt>=2.8",
    "boto3>=1.34",
    "python-dotenv>=1.0",
    "Pillow>=10.0",
    "beautifulsoup4>=4.12",
    "python-ulid>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]
